requirements.txt
streamlit
numpy
pillow
keras==2.15.0
h5py